org - move control utils to shared dir
