+++ src/controls/editor-mode/filters.js
@@ -4,7 +4,7 @@ import { Icons, makeStyles } from '@eqworks/lumen-labs'
 
 import { useStoreState, useStoreActions } from '../../store'
 import WidgetControlCard from '../shared/components/widget-control-card'
-import { renderSection, renderRow } from './util'
+import { renderSection, renderRow } from '../shared/util'
 import PluralLinkedSelect from '../../components/plural-linked-select'
 import RangeFilter from './components/range-filter'
 import CustomSelect from '../../components/custom-select'
