+++ src/controls/editor-mode/right-sidebar.js
@@ -9,7 +9,7 @@ import types from '../../constants/types'
 import CustomSelect from '../../components/custom-select'
 import XYSelect from '../../components/xy-select'
 import ColorSchemeControls from './components/color-scheme-controls'
-import { renderItem, renderSection, renderRow, renderToggle } from './util'
+import { renderItem, renderSection, renderRow, renderToggle } from '../shared/util'
 import UniqueOptionControls from './components/unique-option-controls'
 import EditorSidebarBase from './sidebar-base'
 import Filters from './filters'
