+++ src/controls/shared/domain-controls.js
@@ -5,7 +5,7 @@ import { Chip } from '@eqworks/lumen-labs'
 import { useStoreState, useStoreActions } from '../../store'
 import CustomSelect from '../../components/custom-select'
 import WidgetControlCard from '../shared/components/widget-control-card'
-import { renderRow, renderSection } from '../editor-mode/util'
+import { renderRow, renderSection } from './util'
 import typeInfo from '../../constants/type-info'
 import types from '../../constants/types'
 import CustomRadio from '../../components/custom-radio'
